Remove DistributeTest.java class as it is not in used yet. Moved DistributeUtilTest.java to util folder
