--- src/test/java/seedu/address/model/distribute/DistributeTest.java
@@ -1,7 +1,0 @@
-package seedu.address.model.distribute;
-
-public class DistributeTest {
-
-
-
-}
